requirements.txt
black
tqdm
h5py
polars
scikit-learn
matplotlib
gdown
seaborn
pyyaml
mypy
uniplot
pytorch_forecasting
wandb
sphinx
sphinx-rtd-theme
flask
jsonschema
opencv-python
transformers
allennlp